@@ -3,9 +3,9 @@ description: A Library app that makes it easy to keep track of your favorite boo
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 environment:
-  sdk: '>=2.19.0 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
-isar_version: &isar_version 3.0.5 # define the version to be used
+isar_version: &isar_version 3.1.0+1 # define the version to be used
 
 dependencies:
   flutter:
@@ -18,7 +18,7 @@ dependencies:
   flutter_bloc: ^8.1.2
   equatable: ^2.0.5
   get_it: ^7.2.0
-  go_router: ^6.5.2
+  go_router: ^7.0.1
   books_finder: ^4.3.0
   firebase_auth: ^4.4.0
   firebase_core: ^2.9.0
@@ -29,7 +29,7 @@ dependencies:
   cached_network_image: ^3.2.3
   extended_image:
   logger: ^1.3.0
-  intl: ^0.17.0
+  intl: ^0.18.0
   url_launcher: ^6.1.10
   flutter_native_splash: ^2.2.19
   hydrated_bloc: ^9.1.0
