@@ -1,6 +1,5 @@
 import 'dart:math';
 
-import 'package:books_genie/domain/book/base/entities/base_book.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:books_genie/presentation/shared_widgets/fade_slide_transition.dart';
@@ -17,11 +16,8 @@ class CustomAppbar extends StatelessWidget {
   final Widget? title;
   final String? titleText;
   final double expandedHeight;
-  final void Function(String?)? onSearch;
-  final String searchHint;
-  final bool searchEnabled;
-  final List<BaseBook> searchResultItems;
-  final bool searchInProgress;
+
+  final Widget? searchBox;
 
   const CustomAppbar({
     super.key,
@@ -30,13 +26,9 @@ class CustomAppbar extends StatelessWidget {
     this.searchBarSlideAnimation,
     this.welcomeMessageSlideAnimation,
     this.welcomeMessageFadeAnimation,
-    this.onSearch,
-    this.searchEnabled = true,
     this.title,
     this.titleText,
-    this.searchResultItems = const [],
-    this.searchHint = 'Search...',
-    this.searchInProgress = false,
+    this.searchBox,
     required this.animateComponentsOnInit,
     required this.scrollController,
   }) : assert(
@@ -126,51 +118,31 @@ class CustomAppbar extends StatelessWidget {
               child:
                   _Title(scrollController, title: title, titleText: titleText),
             ),
-            title: CustomAnimatedBuilder(
-              animation: scrollController,
-              animatedChildBuilder: (BuildContext context, Widget child) {
-                final c1 = expandedHeight /
-                    (min(expandedHeight, scrollController.offset));
-                final position = 6.0 * min(2.0, c1 - 1.0);
-                return Positioned(
-                  left: position,
-                  right: position,
-                  bottom: position,
-                  child: child,
-                );
-              },
-              child: FadeSlideTransition(
-                slideAnimation: searchBarSlideAnimation,
-                fadeAnimation: searchBarFadeAnimation,
-                animate: animateComponentsOnInit,
-                child: SizedBox(
-                  height: 54,
-                  child: CupertinoSearchTextField(
-                    enabled: searchEnabled,
-                    prefixInsets: const EdgeInsets.fromLTRB(10, 0, 5, 0),
-                    padding: const EdgeInsetsDirectional.fromSTEB(8, 4, 8, 4),
-                    placeholder: searchHint,
-                    placeholderStyle: context.textTheme.bodyMedium?.copyWith(
-                      color: context.isDarkMode
-                          ? Colors.white70
-                          : context.colorScheme.onBackground.withOpacity(.8),
-                    ),
-                    onChanged: onSearch,
-                    onSubmitted: onSearch,
-                    decoration: BoxDecoration(
-                      boxShadow: [
-                        BoxShadow(
-                          blurRadius: 8,
-                          color:
-                              context.colorScheme.onBackground.withOpacity(.1),
-                        )
-                      ],
-                      color: context.colorScheme.background,
+            title: searchBox == null
+                ? null
+                : CustomAnimatedBuilder(
+                    animation: scrollController,
+                    animatedChildBuilder: (BuildContext context, Widget child) {
+                      final c1 = expandedHeight /
+                          (min(expandedHeight, scrollController.offset));
+                      final position = 6.0 * min(2.0, c1 - 1.0);
+                      return Positioned(
+                        left: position,
+                        right: position,
+                        bottom: position,
+                        child: child,
+                      );
+                    },
+                    child: FadeSlideTransition(
+                      slideAnimation: searchBarSlideAnimation,
+                      fadeAnimation: searchBarFadeAnimation,
+                      animate: animateComponentsOnInit,
+                      child: SizedBox(
+                        height: 54,
+                        child: searchBox,
+                      ),
                     ),
                   ),
-                ),
-              ),
-            ),
           ),
         ],
       ),
