@@ -7,7 +7,7 @@ part of 'isar_user.dart';
 // **************************************************************************
 
 // coverage:ignore-file
-// ignore_for_file: duplicate_ignore, non_constant_identifier_names, constant_identifier_names, invalid_use_of_protected_member, unnecessary_cast, prefer_const_constructors, lines_longer_than_80_chars, require_trailing_commas, inference_failure_on_function_invocation, unnecessary_parenthesis, unnecessary_raw_strings, unnecessary_null_checks, join_return_with_assignment, prefer_final_locals, avoid_js_rounded_ints, avoid_positional_boolean_parameters
+// ignore_for_file: duplicate_ignore, non_constant_identifier_names, constant_identifier_names, invalid_use_of_protected_member, unnecessary_cast, prefer_const_constructors, lines_longer_than_80_chars, require_trailing_commas, inference_failure_on_function_invocation, unnecessary_parenthesis, unnecessary_raw_strings, unnecessary_null_checks, join_return_with_assignment, prefer_final_locals, avoid_js_rounded_ints, avoid_positional_boolean_parameters, always_specify_types
 
 extension GetIsarUserCollection on Isar {
   IsarCollection<IsarUser> get isarUsers => this.collection();
@@ -109,7 +109,7 @@ const IsarUserSchema = CollectionSchema(
   getId: _isarUserGetId,
   getLinks: _isarUserGetLinks,
   attach: _isarUserAttach,
-  version: '3.0.5',
+  version: '3.1.0+1',
 );
 
 int _isarUserEstimateSize(
